PFDR-195 - Move storage classes into Chipmunk module
