+++ lib/chipmunk.rb
@@ -10,7 +10,6 @@ require_relative "chipmunk/validatable"
 
 require_relative "chipmunk/bag"
 require_relative "chipmunk/resolvers"
-require_relative "incoming_storage"
-require_relative "package_storage"
-
-require_relative "volume"
+require_relative "chipmunk/incoming_storage"
+require_relative "chipmunk/package_storage"
+require_relative "chipmunk/volume"
