+++ spec/support/contexts/with_test_volume.rb
@@ -2,7 +2,7 @@
 
 RSpec.shared_context "with test volume" do
   before(:each) do
-    volume = Volume.new(name: "test", package_type: Chipmunk::Bag, root_path: Rails.root/"spec/support/fixtures")
+    volume = Chipmunk::Volume.new(name: "test", package_type: Chipmunk::Bag, root_path: Rails.root/"spec/support/fixtures")
     allow(Services.storage).to receive(:for) do |package|
       volume.get(package.storage_path)
     end
